% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_data.R
\name{hydat_flow_monthly}
\alias{hydat_flow_monthly}
\alias{hydat_level_monthly}
\alias{hydat_sed_monthly}
\alias{hydat_flow_daily}
\alias{hydat_level_daily}
\alias{hydat_sed_daily}
\title{Retreive HYDAT Data}
\usage{
hydat_flow_monthly(stationid, year = NULL, month = 1:12, db = hydat_get_db())

hydat_level_monthly(stationid, year = NULL, month = 1:12, db = hydat_get_db())

hydat_sed_monthly(stationid, year = NULL, month = 1:12, db = hydat_get_db())

hydat_flow_daily(
  stationid,
  year = NULL,
  month = 1:12,
  day = 1:31,
  db = hydat_get_db()
)

hydat_level_daily(
  stationid,
  year = NULL,
  month = 1:12,
  day = 1:31,
  db = hydat_get_db()
)

hydat_sed_daily(
  stationid,
  year = NULL,
  month = 1:12,
  day = 1:31,
  db = hydat_get_db()
)
}
\arguments{
\item{stationid}{The station identifier(s) for which data should be loaded}

\item{year}{The year or range of years for which data should be loaded, or NULL for all}

\item{month}{The month or range of months for which data should be loaded}

\item{db}{The hydat database object to use (you will want to use
\link{hydat_load} before using these functions)}

\item{day}{The day or range of days for which data should be loaded}
}
\value{
A data.frame (tibble, actually) of observations
}
\description{
Retreive HYDAT Data
}
\examples{
hydat_load_test_db()
hydat_flow_monthly("01AD001")
hydat_flow_daily("01AD001")
hydat_level_monthly("01AD003")
hydat_level_daily("01AD003")
hydat_sed_monthly("01AF006")
hydat_sed_daily("01AF006")

}
