% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydat_date_seq}
\alias{hydat_date_seq}
\alias{hydat_month_seq}
\title{Create A Continuous Date Sequnce}
\usage{
hydat_date_seq(year, month = 1:12, day = 1:31)

hydat_month_seq(year, month = 1:12)
}
\arguments{
\item{year}{A vector of years}

\item{month}{A vector of months (default, 1:12)}

\item{day}{A vector of days (default, 1:31)}
}
\value{
A data.frame with columns YEAR, MONTH, DAY, and DATE
}
\description{
Create A Continuous Date Sequnce
}
\examples{
hydat_date_seq(2000:2011)
hydat_month_seq(2000:2011)

}
