% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_db.R
\name{hydat_download}
\alias{hydat_download}
\title{Download the Hydat Database from Environment Canada}
\usage{
hydat_download(destination = ".", overwrite = NA, quiet = FALSE)
}
\arguments{
\item{destination}{A filename or directory}

\item{overwrite}{Should the destination file be overwritten? Use NA to
skip downloading a previously downloaded file (the default).}

\item{quiet}{If TRUE, download status messages will be suppressed.}
}
\value{
The filename of the downloaded file.
}
\description{
Downloads the latest version of the Hydat Database from
\url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/}. This
package uses the zipped SQLite version of the database. The download is about
200 MB, and usually takes 5-10 minutes to download.
}
\examples{
\dontrun{
# this usually takes around 10 minutes to run
hydat_download()
}

}
