% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_download.R
\name{hydat_extract}
\alias{hydat_extract}
\title{Extract a Zipped Hydat Database}
\usage{
hydat_extract(source = ".", destination = ".", overwrite = NA)
}
\arguments{
\item{source}{The directory that contains the hydat database or the hydat zip file.}

\item{destination}{The directory that the output zip file should be unzipped to.}

\item{overwrite}{Should existing files be overwritten? (use NA to skip a previously
extracted file).}
}
\value{
The filename of the extracted sqlite database file
}
\description{
A zipped version of the Hydat database can be obtained using
\link{hydat_download}. The extracted database is more than 1 GB in size.
If multiple downloaded archives exist in the source directory, the latest one is
used with a message.
}
\examples{
\dontrun {
# hydat_download() takes about 10 minutes; hydat_extract() takes about 2 minutes
hydat_download()
hydat_extract()
}
}
