% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_stations.R
\name{hydat_find_stations}
\alias{hydat_find_stations}
\title{Find HYDAT Data Stations}
\usage{
hydat_find_stations(x, year = NULL, limit = 10, db = hydat_get_db())
}
\arguments{
\item{x}{A character vector of a place name (geocoded by \link[prettymapr]{geocode}),
or a numeric vector of length 2 (lon, lat).}

\item{year}{Optional range of years that must be included in the station data}

\item{limit}{This number of closest stations will be returned}

\item{db}{The hydat database object to use (you will want to use
\link{hydat_load} before using these functions)}
}
\value{
A data.frame (tibble) of station information
}
\description{
Find HYDAT Data Stations
}
\examples{
hydat_load_test_db()
hydat_find_stations(c(-70.08, 46.55))
hydat_find_stations(c(-70.08, 46.55), year = 1988:1991)

hydat_find_stations("Daaquam, QC")
hydat_find_stations("Daaquam, QC", year = 1988:1991)


}
