% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_db.R
\name{hydat_list_files}
\alias{hydat_list_files}
\title{List Hydat files}
\usage{
hydat_list_files(source = ".", ext = ".db", full.names = TRUE,
  recursive = FALSE)
}
\arguments{
\item{source}{The directory to search for Hydat files}

\item{ext}{The extension of the file}

\item{full.names}{Should full names be used?}

\item{recursive}{Should directories be searched recursively?}
}
\value{
A character vector of filenames
}
\description{
List Hydat files with the specified extension.
}
\examples{
hydat_list_files()

}
