% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_db.R
\name{hydat_load}
\alias{hydat_load}
\title{Load a previously extracted Hydat Database}
\usage{
hydat_load(source = ".", set = TRUE)
}
\arguments{
\item{source}{The directory where the Hydat database was extracted, or the
database file.}

\item{set}{Pass FALSE to skip setting the internal package reference to the loaded
database (advanced).}
}
\value{
An object extending \link[dplyr]{src_sqlite}, invisibly
}
\description{
This function uses \link[dplyr]{src_sqlite} to load the sqlite database
extracted by \link{hydat_extract} and downloaded using \link{hydat_download}.
}
\examples{
\dontrun{
# hydat_download() takes about 10 minutes; hydat_extract() takes about 2 minutes
hydat_download()
hydat_extract()
hydat_db <- hydat_load()
dplyr::src_tbls(hydat_db)
}
}
