% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_db.R
\name{hydat_set_db}
\alias{hydat_set_db}
\alias{hydat_get_db}
\alias{hydat_load_test_db}
\title{Set/get the internal pointer to the Hydat database}
\usage{
hydat_set_db(x)

hydat_get_db()

hydat_load_test_db(set = TRUE)
}
\arguments{
\item{x}{A hydat datbase loaded using \link{hydat_load}.}

\item{set}{Should the temporary database be set as the internal
database?}
}
\value{
The hydat database or the previous hydat database.
}
\description{
The functions in this package require the Hydat database to be loaded
using \link{hydat_load} or hydat_set_db(), or for one to be explicitly
passed to each function. Generally this is done using \link{hydat_load}
with \code{set = TRUE}.
}
\examples{
\dontrun{
hydat_download()
hydat_extract()
hydat_db <- hydat_load() # this calls hydat_set_db()
hydat_set_db(hydat_db) # this is redundant unless set = FALSE in hydat_load()
}

}
