% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_stations.R
\name{hydat_station_info}
\alias{hydat_station_info}
\title{Get Detailed Station Info}
\usage{
hydat_station_info(stationid = NULL, db = hydat_get_db())
}
\arguments{
\item{stationid}{A station identifier, or NULL for all stations}

\item{db}{db The hydat database object to use (you will want to use
\link{hydat_load} before using these functions)}
}
\value{
A data.frame (tibble) with one row per station
}
\description{
Get Detailed Station Info
}
\examples{
hydat_load_test_db()
hydat_station_info()
hydat_station_info("01AD012")
as.list(hydat_station_info("01AD012"))

}
