% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_db.R
\name{hydat_tbl}
\alias{hydat_tbl}
\title{Low-level access to HYDAT database tables}
\usage{
hydat_tbl(table, db = hydat_get_db())
}
\arguments{
\item{table}{The name of the table. As of this writing, the list of table
names that can be used includes AGENCY_LIST, ANNUAL_INSTANT_PEAKS,
ANNUAL_STATISTICS, CONCENTRATION_SYMBOLS, DATA_SYMBOLS, DATA_TYPES,
DATUM_LIST, DLY_FLOWS, DLY_LEVELS, MEASUREMENT_CODES, OPERATION_CODES,
PEAK_CODES, PRECISION_CODES, REGIONAL_OFFICE_LIST, SAMPLE_REMARK_CODES,
SED_DATA_TYPES, SED_DLY_LOADS, SED_DLY_SUSCON, SED_SAMPLES,
SED_SAMPLES_PSD, SED_VERTICAL_LOCATION, SED_VERTICAL_SYMBOLS, STATIONS,
STN_DATA_COLLECTION, STN_DATA_RANGE, STN_DATUM_CONVERSION,
STN_DATUM_UNRELATED, STN_OPERATION_SCHEDULE, STN_REGULATION, STN_REMARKS,
STN_REMARK_CODES, STN_STATUS_CODES, and VERSION}

\item{db}{The hydat database object to use (you will want to use
\link{hydat_load} before using these functions)}
}
\value{
A tbl_SQLiteConnection object (use dplyr to work with these objects)
}
\description{
This function gives low-level access to the underlying HYDAT tables used by
other functions. Many of these tables are too large to load into memory,
so it is best to use dplyr to \link[dplyr]{filter} them before using
\link[dplyr]{collect} to read them into memory.
}
\examples{
hydat_load_test_db()
hydat_tbl("STATIONS")

}
