% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_package.R
\docType{package}
\name{hydatr-package}
\title{An R Interface to the Environment Canada HYDAT database}
\description{
The HYDAT database contains over 1 GB of hydrological
observation data collected by Environment Canada. The database
is freely available from the Environment Canada website, however
exctracting data can be difficult. This package provides tools to
easily find and extract hydrological data from the HYDAT database.
}
