% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_download.R
\name{is_hydat}
\alias{is_hydat}
\title{Test if an object was created using hydat_load}
\usage{
is_hydat(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
TRUE if the object inherits src_hydat, FALSE otherwise
}
\description{
Test if an object was created using hydat_load
}
\examples{
\dontrun {
hydat_db <- hydat_load()
is_hydat(hydat_db)
}

}
